Add collapsing toolbar
The toolbar is a custom implementation of the material top app bar. It gradually collapses and shrinks the title size upon scroll. This behavior is similar to the Coordinator layout behavior from the XML View world.

The code is structured in the way that the current behavior (exit until collapsed) can be easily replaced with another one in the future.

Unfortunately, compose lacks an official implementation for this at the moment. There is only experimental one with fade in/fade out, which also has an issue with cropping the title when it's two lines in height.

The code for the collapsing toolbar behavior was taken from the following repository: https://github.com/pencelab/CollapsingToolbarInCompose. Certain modifications were applied.