Change logic of updating local change list versions
When the data is prepopulated, update change list versions to the latest ones from the network. However, if the change list is empty, update to 0.

Because it would be good to somehow indicate that the data is up to date. The value of -1 doesn't suggest that. So update to 0 instead.