Implement saving grocery list order positions in local database
The user can reorder items by dragging them within the list. New item positions are saved to the database only when the item is moved and the drag is fully finished (at this moment the item touch helper callback's clearView is called).

When the flow of dashboard items is fired off, the AndroidView's update callback is called and the data in the adapter gets updated. Then DiffUtil calculates differences between new and old list, and notifies the recyclerview about the changes in a performant way.