Configure preservation of last opened list
The app now opens the grocery list that the user visited last time. If the user didn't open any lists or the list is not there, opens the dashboard screen.

The last opened id is stored in preferences data store. Then it's retrieved in MainActivityViewModel and passed to GroceryGeniusNavHost that determines the start navigation destination.

The id stored in the data store may become outdated because it's not connected to changes in the database in any way. However, the validity of the data is handled in UserPreferencesRepository, which checks whether the list with that id exists, and returns null if that's not the case.

Additionally made some fixes in database layer.