Configure Retrofit and fetch predefined data from GitHub repo
From now on, the predefined data such as products, categories, and icons is fetched from the GroceryGenius repository on GitHub. GitHub is like a cloud provider for our files.

This solution isn't very good but, for now, that we don't have a running backend, it's better than bundling these things into the APK. So going for a scalable approach here.

 It's also convenient as the data can be easily updated with a git commit and the update will be reflected on all devices without the need to ship another version.