Add OnboardingSyncScreen and show it while the database is getting prepopulated
This screen is shown only when the local database is empty and is never shown when the predefined data gets loaded. The purpose of this screen is to ask users to connect to the internet in such an edge case when the app is launched at the first time and the device is not connected to the internet. Also, it informs the user if the network request fails.