Improve deriving custom product and fix wrong groceries color in the bottom sheet
For example there is a product "sauce" in the database, but the user searches for "tomato sauce", which is not there. In this case, the custom product "tomato sauce" will get the icon and category of the existing product "sauce". Previously, only icon was derived.

Additionally fixed wrong groceries color in the AddGroceryBottomSheetContent.